toStrings y metodos adicionales

+++ src/com/educationallab/console/Persona.java
@@ -6,13 +6,13 @@ import java.io.Serializable;
 public class Persona implements Serializable {
 
     private  static  final  long serialVersionUID = 1L;
-    private Integer id;
+    private Double id;
     private String nombres;
     private String apellidos;
     private String email;
 
 
-    public Persona(Integer id, String nombres, String apellidos, String email){
+    public Persona(Double id, String nombres, String apellidos, String email){
         this.id = id;
         this.nombres = nombres;
         this.apellidos = apellidos;
@@ -20,7 +20,7 @@ public class Persona implements Serializable {
     }
 
 
-    public Integer getId() {
+    public Double getId() {
         return id;
     }
